Ensure all logic is in craftax.h and clean up function names.
