+++ ocean/craftax_clean/craftax_clean.h
@@ -1755,8 +1755,78 @@ void update_log_state(Craftax* env) {
     }
 }
 
-#define CRAFTAX_CLEAN_IN_HEADER
-#include "craftax_clean.c"
+static int key_to_action(void) {
+    static const int map[][2] = {
+        {KEY_Q, ACTION_NOOP},
+        {KEY_W, ACTION_UP},
+        {KEY_UP, ACTION_UP},
+        {KEY_D, ACTION_RIGHT},
+        {KEY_RIGHT, ACTION_RIGHT},
+        {KEY_S, ACTION_DOWN},
+        {KEY_DOWN, ACTION_DOWN},
+        {KEY_A, ACTION_LEFT},
+        {KEY_LEFT, ACTION_LEFT},
+        {KEY_SPACE, ACTION_DO},
+        {KEY_ONE, ACTION_MAKE_WOOD_PICKAXE},
+        {KEY_TWO, ACTION_MAKE_STONE_PICKAXE},
+        {KEY_THREE, ACTION_MAKE_IRON_PICKAXE},
+        {KEY_FOUR, ACTION_MAKE_DIAMOND_PICKAXE},
+        {KEY_FIVE, ACTION_MAKE_WOOD_SWORD},
+        {KEY_SIX, ACTION_MAKE_STONE_SWORD},
+        {KEY_SEVEN, ACTION_MAKE_IRON_SWORD},
+        {KEY_EIGHT, ACTION_MAKE_DIAMOND_SWORD},
+        {KEY_T, ACTION_PLACE_TABLE},
+        {KEY_TAB, ACTION_SLEEP},
+        {KEY_R, ACTION_PLACE_STONE},
+        {KEY_F, ACTION_PLACE_FURNACE},
+        {KEY_P, ACTION_PLACE_PLANT},
+        {KEY_E, ACTION_REST},
+        {KEY_COMMA, ACTION_ASCEND},
+        {KEY_PERIOD, ACTION_DESCEND},
+        {KEY_Y, ACTION_MAKE_IRON_ARMOUR},
+        {KEY_U, ACTION_MAKE_DIAMOND_ARMOUR},
+        {KEY_I, ACTION_SHOOT_ARROW},
+        {KEY_O, ACTION_MAKE_ARROW},
+        {KEY_G, ACTION_CAST_FIREBALL},
+        {KEY_H, ACTION_CAST_ICEBALL},
+        {KEY_J, ACTION_PLACE_TORCH},
+        {KEY_Z, ACTION_DRINK_POTION_RED},
+        {KEY_X, ACTION_DRINK_POTION_GREEN},
+        {KEY_C, ACTION_DRINK_POTION_BLUE},
+        {KEY_V, ACTION_DRINK_POTION_PINK},
+        {KEY_B, ACTION_DRINK_POTION_CYAN},
+        {KEY_N, ACTION_DRINK_POTION_YELLOW},
+        {KEY_M, ACTION_READ_BOOK},
+        {KEY_K, ACTION_ENCHANT_SWORD},
+        {KEY_L, ACTION_ENCHANT_ARMOUR},
+        {KEY_LEFT_BRACKET, ACTION_MAKE_TORCH},
+        {KEY_RIGHT_BRACKET, ACTION_LEVEL_UP_DEXTERITY},
+        {KEY_MINUS, ACTION_LEVEL_UP_STRENGTH},
+        {KEY_EQUAL, ACTION_LEVEL_UP_INTELLIGENCE},
+        {KEY_SEMICOLON, ACTION_ENCHANT_BOW},
+    };
+    for (int i = 0; i < (int)(sizeof(map) / sizeof(map[0])); i++) {
+        if (IsKeyPressed(map[i][0])) {
+            return map[i][1];
+        }
+    }
+    return -1;
+}
+
+// Shift + action-panel key. 1 = applied, 0 = policy, -1 = skip tick.
+static int got_human_input(Craftax* env) {
+    int shift = IsKeyDown(KEY_LEFT_SHIFT) || IsKeyDown(KEY_RIGHT_SHIFT);
+    if (!IsWindowReady() || !shift
+        || env->state.is_sleeping || env->state.is_resting) {
+        return 0;
+    }
+    int action = key_to_action();
+    if (action < 0) {
+        return -1;
+    }
+    env->agents[0].actions[0] = (float)action;
+    return 1;
+}
 
 void puf_reset(Craftax* env) {
     env->agents[0].rewards[0] = 0.0f;
@@ -1785,7 +1855,7 @@ void puf_reset(Craftax* env) {
 }
 
 void puf_step(Craftax* env) {
-    if (craftax_clean_human_controls(env) < 0) {
+    if (got_human_input(env) < 0) {
         return;
     }
     env->agents[0].rewards[0] = 0.0f;
@@ -2912,7 +2982,7 @@ static void draw_inv_slot(int tex_id, int overlay, int x, int y) {
     }
 }
 
-static const char* craftax_clean_action_names[ATN_DIM] = {
+static const char* action_names[ATN_DIM] = {
     "NOOP", "LEFT", "RIGHT", "UP", "DOWN", "DO", "SLEEP",
     "PLACE_STONE", "PLACE_TABLE", "PLACE_FURNACE", "PLACE_PLANT",
     "MAKE_WOOD_PICKAXE", "MAKE_STONE_PICKAXE", "MAKE_IRON_PICKAXE",
@@ -2928,7 +2998,7 @@ static const char* craftax_clean_action_names[ATN_DIM] = {
     "ENCHANT_BOW",
 };
 
-static const char* craftax_clean_action_keys[ATN_DIM] = {
+static const char* action_keys[ATN_DIM] = {
     "Q", "A", "D", "W", "S", "Space", "Tab",
     "R", "T", "F", "P",
     "1", "2", "3",
@@ -2943,7 +3013,7 @@ static const char* craftax_clean_action_keys[ATN_DIM] = {
     ";",
 };
 
-static const char* craftax_clean_ach_names[NUM_ACHIEVEMENTS] = {
+static const char* ach_names[NUM_ACHIEVEMENTS] = {
     "Collect Wood", "Place Table", "Eat Cow", "Collect Sapling", "Collect Drink",
     "Make Wood Pickaxe", "Make Wood Sword", "Place Plant", "Defeat Zombie",
     "Collect Stone", "Place Stone", "Eat Plant", "Defeat Skeleton",
@@ -2997,7 +3067,7 @@ void puf_render(Craftax* env) {
             }
         }
         if (textures.id == 0) {
-            fprintf(stderr, "craftax_clean: textures.png not found in resources/craftax\n");
+            fprintf(stderr, "craftax textures.png not found in resources/craftax\n");
             exit(1);
         }
         SetTextureFilter(textures, TEXTURE_FILTER_POINT);
@@ -3006,7 +3076,7 @@ void puf_render(Craftax* env) {
     if (IsKeyDown(KEY_ESCAPE)) {
         exit(0);
     }
-    craftax_clean_human_controls(env);
+    got_human_input(env);
 
     int level = clampi(env->state.player_level, 0, NUM_LEVELS - 1);
     int player_row = clampi(env->state.player_position[0], 0, MAP_SIZE - 1);
@@ -3268,8 +3338,8 @@ void puf_render(Craftax* env) {
         }
         Color text_color = selected ? BLACK
             : (legal ? (Color){220, 230, 230, 255} : (Color){80, 90, 90, 255});
-        DrawText(craftax_clean_action_keys[action], panel_x + 12, y, 10, text_color);
-        DrawText(TextFormat("%02d %s", action, craftax_clean_action_names[action]),
+        DrawText(action_keys[action], panel_x + 12, y, 10, text_color);
+        DrawText(TextFormat("%02d %s", action, action_names[action]),
             panel_x + 78, y, 10, text_color);
     }
 
@@ -3288,7 +3358,7 @@ void puf_render(Craftax* env) {
             DrawRectangle(0, y, ACH_PANEL_W, ach_row, (Color){46, 180, 80, 255});
         }
         DrawText(
-            craftax_clean_ach_names[i],
+            ach_names[i],
             6,
             y + (ach_row > 10 ? 1 : 0),
             10,
